# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NexDL Download Manager API
servers:
  - url: /api
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List all saved accounts
      responses:
        "200":
          description: Account list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Add a new account with cookies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get account details
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        "200":
          description: Account detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
    patch:
      operationId: updateAccount
      tags: [accounts]
      summary: Update account label or cookies
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUpdate"
      responses:
        "200":
          description: Account updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Remove an account
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        "204":
          description: Removed

  # ── Scheduler ─────────────────────────────────────────────────────────────
  /scheduler/entries:
    get:
      operationId: listScheduleEntries
      tags: [scheduler]
      summary: List all schedule entries
      responses:
        "200":
          description: Schedule list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleEntry"
    post:
      operationId: createScheduleEntry
      tags: [scheduler]
      summary: Create a new schedule entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleInput"
      responses:
        "201":
          description: Entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleEntry"
  /scheduler/entries/{id}:
    patch:
      operationId: updateScheduleEntry
      tags: [scheduler]
      summary: Update a schedule entry
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleUpdate"
      responses:
        "200":
          description: Entry updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleEntry"
    delete:
      operationId: deleteScheduleEntry
      tags: [scheduler]
      summary: Remove a schedule entry
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        "204":
          description: Removed

  # ── Settings ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Account:
      type: object
      required: [id, label, site, cookieCount, createdAt]
      properties:
        id:
          type: string
        label:
          type: string
        site:
          type: string
        username:
          type: ["string", "null"]
        cookieCount:
          type: integer
        createdAt:
          type: string
    AccountInput:
      type: object
      required: [label, site]
      properties:
        label:
          type: string
        site:
          type: string
        username:
          type: ["string", "null"]
        cookies:
          type: ["string", "null"]
          description: Netscape or JSON cookie string
    AccountUpdate:
      type: object
      properties:
        label:
          type: string
        username:
          type: ["string", "null"]
        cookies:
          type: ["string", "null"]

    # ── Scheduler ─────────────────────────────────────────────────────────
    ScheduleEntry:
      type: object
      required: [id, name, triggerType, action, enabled]
      properties:
        id:
          type: string
        name:
          type: string
        triggerType:
          type: string
          enum: [once, interval, cron, on_startup]
        triggerValue:
          type: ["string", "null"]
          description: ISO datetime for once, seconds for interval, cron expression for cron
        action:
          type: string
          enum: [download, run_addon, cleanup]
        actionTarget:
          type: ["string", "null"]
          description: URL for download, addon name for run_addon
        enabled:
          type: boolean
        lastRun:
          type: ["string", "null"]
        nextRun:
          type: ["string", "null"]
    ScheduleInput:
      type: object
      required: [name, triggerType, action]
      properties:
        name:
          type: string
        triggerType:
          type: string
          enum: [once, interval, cron, on_startup]
        triggerValue:
          type: ["string", "null"]
        action:
          type: string
          enum: [download, run_addon, cleanup]
        actionTarget:
          type: ["string", "null"]
        enabled:
          type: boolean
    ScheduleUpdate:
      type: object
      properties:
        name:
          type: string
        triggerType:
          type: string
          enum: [once, interval, cron, on_startup]
        triggerValue:
          type: ["string", "null"]
        action:
          type: string
          enum: [download, run_addon, cleanup]
        actionTarget:
          type: ["string", "null"]
        enabled:
          type: boolean

    # ── Settings ──────────────────────────────────────────────────────────
...